expense_reimbursement_system/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
pytesseract==0.3.10
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==23.5
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
